platform net check, working on local storage for chats

+++ lib/auth/controllers/auth_cubit_impl.dart
@@ -6,6 +6,7 @@ import 'package:chat_flutter_firebase/auth/controllers/auth_cubit.dart';
 import 'package:chat_flutter_firebase/auth/controllers/auth_state.dart';
 import 'package:chat_flutter_firebase/auth/services/auth_service.dart';
 import 'package:chat_flutter_firebase/common/app_text.dart';
+import 'package:chat_flutter_firebase/connectivity/network_connectivity.dart';
 import 'package:chat_flutter_firebase/local_storage/local_models/local_user_info.dart';
 import 'package:chat_flutter_firebase/local_storage/services/local_storage_service.dart';
 import 'package:chat_flutter_firebase/rest_network/network_service.dart';
@@ -16,15 +17,17 @@ class AuthCubitImpl extends Cubit<AuthState> implements AuthCubit {
   AuthCubitImpl({
     required AuthService authService,
     required NetworkService networkService,
-    required LocalStorageService localStorageService
+    required LocalStorageService localStorageService,
+    required NetworkConnectivity networkConnectivity,
   })
       : _auth = authService,
-        _network = networkService,
+        _networkService = networkService,
         _localStorage = localStorageService,
+        _networkConnectivity = networkConnectivity,
         super(const AuthState(status: AuthStatus.signedOut))
   {
     userSubscription = _auth.authStateChanges().listen((user) {
-      log(user.toString());
+      log('auth state subscription: ${user.toString()}');
       user == null
           ? emit(state.copyWith(status: AuthStatus.signedOut, user: null))
           : emit(state.copyWith(status: AuthStatus.signedIn, user: user));
@@ -32,14 +35,18 @@ class AuthCubitImpl extends Cubit<AuthState> implements AuthCubit {
   }
 
   final AuthService _auth;
-  final NetworkService _network;
+  final NetworkService _networkService;
   final LocalStorageService _localStorage;
+  final NetworkConnectivity _networkConnectivity;
   StreamSubscription<UserInfo?>? userSubscription;
 
   @override
   Future<void> createUserByEmailAndPassword(
       String email, String password, String? username) async {
     try {
+      if (!await _checkNetworkConnection()) {
+        return;
+      }
       emit(state.copyWith(status: AuthStatus.loading));
       await _auth.createUserByEmailAndPassword(email, password);
       if (username != null) {
@@ -60,6 +67,9 @@ class AuthCubitImpl extends Cubit<AuthState> implements AuthCubit {
   @override
   Future<void> signInByEmailAndPassword(String email, String password) async {
     try {
+      if (!await _checkNetworkConnection()) {
+        return;
+      }
       emit(state.copyWith(status: AuthStatus.loading));
       await _auth.signInByEmailAndPassword(email, password);
       await _addUserIfAbsent();
@@ -77,6 +87,9 @@ class AuthCubitImpl extends Cubit<AuthState> implements AuthCubit {
   @override
   Future<void> signInByGoogle() async {
     try {
+      if (!await _checkNetworkConnection()) {
+        return;
+      }
       emit(state.copyWith(status: AuthStatus.loading));
       await _auth.signInByGoogle();
       await _addUserIfAbsent();
@@ -93,6 +106,9 @@ class AuthCubitImpl extends Cubit<AuthState> implements AuthCubit {
   @override
   Future<void> signOut() async {
     try {
+      if (!await _checkNetworkConnection()) {
+        return;
+      }
       await _auth.signOut();
       await _localStorage.deleteCurrentAppUser();
     } catch (e) {
@@ -104,7 +120,6 @@ class AuthCubitImpl extends Cubit<AuthState> implements AuthCubit {
   }
 
   @override
-  //TODO if no connection => from local storage;
   UserInfo? get user => state.user;
 
   @override
@@ -113,14 +128,23 @@ class AuthCubitImpl extends Cubit<AuthState> implements AuthCubit {
     await super.close();
   }
 
-  Future<void> _addUserIfAbsent() async {
-    !await _network.isUserInDatabase(state.user!)
-        ? await _network.saveUser(state.user!)
-        : null;
-  }
-
   @override
   void setUserFromLocalStorage(LocalUserInfo localUserInfo) {
     emit(state.copyWith(user: UserInfo.fromLocalUserInfo(localUserInfo)));
+  }
+
+  Future<void> _addUserIfAbsent() async {
+    !await _networkService.isUserInDatabase(state.user!)
+        ? await _networkService.saveUser(state.user!)
+        : null;
+  }
+
+  Future<bool> _checkNetworkConnection() async {
+    final result = await _networkConnectivity.checkNetworkConnection();
+    !result
+        ? emit(state.copyWith(
+            status: AuthStatus.error, message: AuthErrorTexts.noConnectionRU))
+        : null;
+    return result;
   }
 }
